Implementing a test that attempts to instantiate a few different ways
